\name{l2path-package}
\alias{l2path-package}
\alias{l2path}
\docType{package}
\title{
\packageTitle{l2path}
}
\description{
\packageDescription{l2path}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{l2path}
\packageIndices{l2path}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{l2path}

Maintainer: \packageMaintainer{l2path}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
Optional links to other man pages, e.g.
}
