% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logistic_GD_path}
\alias{logistic_GD_path}
\title{Solution path of \eqn{\ell_2} regularized logistic regression by gradient descent}
\usage{
logistic_GD_path(X, Y, t_max, alpha_init)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} binary vector}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{alpha_init}{Initial step size specified to start the iteration}
}
\value{
theta Generated solution path

alpha_t_vec Adaptively chosen step size

t_vec Adaptively selected grid points

n_vec Number of gradient steps at each grid point
}
\description{
Generate solution of \eqn{\ell_2} regularized logistic regression by using gradient descent method at each grid point.
Grid points are adaptively selected to save computation while maintaining the high accuracy of solution path.
}
\details{
This function will be called by the main function l2path, with class = "logistic" and method = "GD".
}
