% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{regression_Newton_path_high_dimension_vary}
\alias{regression_Newton_path_high_dimension_vary}
\title{Solution path of ridge regression by Newton method (\eqn{n < p})}
\usage{
regression_Newton_path_high_dimension_vary(X, Y, t_max, alpha_init)
}
\arguments{
\item{X}{A \eqn{n \times p} feature matrix}

\item{Y}{A \eqn{n \times 1} response vector}

\item{t_max}{Range of solution path \eqn{[0, t_{\max}]}}

\item{alpha_init}{Initial step size specified to start the iteration}
}
\value{
theta Generated solution path

alpha_t_vec Adaptively chosen step size

t_vec Adaptively selected grid points
}
\description{
Generate solution of ridge regression by using one-step Newton method at each grid point. 
Woodbury matrix identity is applied to compute the matrix inverse.
Grid points are adaptively selected to save computation while maintaining the high accuracy of solution path.
}
\details{
This function will be called by the main function l2path, with class = "square" and method = "Newton_high".
}
